Add Rust and C demos for testing
This has posed a bit of an issue. Writing the Rust function the proper
way (i.e. not relying on bools/ints being the same thing) breaks the
pattern we're using to look for branches that depend on specific
function calls. We can try to make the pattern checking generic, we
could tailor it to the Rust IR product, or we can alter the Rust source
to produce IR that looks like C.